feat: replace zOrder int with float for precision & add Interactable component
- Updated `zOrder` in all applicable components, structs, and prefabs (`Sprite`, `Rectangle`, `Text`, `Button`, etc.) from `int` to `float` to support more precise rendering depth control.
- Introduced a new `Interactable` component for better UI interaction handling.
- Refactored prefab creation logic in `Button` to ensure proper entity hierarchy (`Interactable`, `Clickable`, background, and text components).
- Adjusted all affected scenes (`PauseScene`, `MainMenuScene`, `BouncingBallScene`) and modules (`UIModule`, `RenderModule`) to integrate refined `zOrder` and `Interactable` features.
- Improved click handling by replacing bounding checks with `Transform` and `Size`-based containment.
- Cleaned up redundant includes and logic in multiple modules to enhance maintainability.

+++ Source/MyGame/Private/GameStates/Gameplay/Scenes/BouncingBallScene.cpp
@@ -33,7 +33,7 @@ void BouncingBallScene::Initialize()
         {.gravity = PhysicsConstants::NO_GRAVITY, .pixelsPerCentimeter = PhysicsConstants::PIXELS_PER_CENTIMETER}
     );
 
-    int zOrder = 0;
+    float zOrder = 0.f;
 
     Prefabs::Rectangle::Create(
         world,
@@ -49,7 +49,7 @@ void BouncingBallScene::Initialize()
     CreateBalls(world, zOrder);
 }
 
-void BouncingBallScene::CreateBalls(const flecs::world& world, int zOrder)
+void BouncingBallScene::CreateBalls(const flecs::world& world, float zOrder)
 {
     ZoneScopedN("BouncingBallScene::CreateBalls");
