+++ Source/MyGame/Private/Modules/UI/Prefabs/Text.h
@@ -23,7 +23,7 @@ struct TextParams
     sf::Vector2f position = {0.f, 0.f};
     sf::Vector2f scale = {1.f, 1.f};
     float rotation = 0.f;
-    int zOrder = 0;
+    float zOrder = 0;
 };
 
 struct Text
